requirements.txt
Django==4.2.3
djangorestframework==3.14.0
scikit-learn==1.3.0
uWSGI==2.0.22
